+++ shared/transaction-modal/tx-stages/icons.tsx
