+++ shared/transaction-modal/tx-stages/iconsStyles.ts
