+++ shared/transaction-modal/tx-stages/tx-stage-fail.tsx
