+++ shared/transaction-modal/tx-stages/tx-stage-permit.tsx
