+++ shared/transaction-modal/tx-stages/tx-stage-sign.tsx
